requirements.txt
streamlit
google-generativeai
pyttsx3
SpeechRecognition
pyaudio
